requirements.txt
# Core dependencies
python-telegram-bot==20.8
deep-translator==1.11.4
python-dotenv==1.0.0
